Refactor: 홈 페이지 로직 분리 및 userName API 연동
